# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CERBERUS Commerce OS API
  version: "4.1-production-readiness"
  description: >
    Çekirdek route-handler sözleşmesi. Uçlar varsayılan olarak HttpOnly
    `cerberus_session` JWT çereziyle korunur. Açık uçlar `/api/auth/login`,
    `/api/health` ve `/api/health/ready` olarak security alanında ayrıca işaretlenir.
    Kimliksiz/geçersiz oturum 401, rol veya mağaza kapsamı ihlali 403, body
    doğrulama hatası 422, aşırı gövde 413 ve beklenmeyen sunucu hatası
    correlationId ile 500 döner. STORE_USER için mağaza kapsamı sunucuda zorlanır.
servers:
  - url: /
security:
  - cookieAuth: []
...
paths:
...
  /api/orders:
    get:
      tags: [orders]
      summary: Sipariş listesi + SQL KPI + sayfalama
      parameters:
        - { name: storeCode, in: query, schema: { type: string } , description: "STORE_USER için yok sayılır (F-11)" }
        - { name: cargoStatus, in: query, schema: { type: string } }
        - { name: pshBatchNo, in: query, schema: { type: string } }
        - { name: search, in: query, description: "Sipariş no, ASIN, MSKU, ürün, marka, tedarikçi ve e-posta içinde arar; ilk 120 karakter kullanılır", schema: { type: string, maxLength: 120 } }
        - { name: page, in: query, schema: { type: integer, default: 1, minimum: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, default: 50, minimum: 1, maximum: 200 } }
      responses:
        "200": { description: "Maskelenmiş orders sayfası, mağaza kapsamlı metadata, filtrelenmiş tüm kümenin kpis değeri ve pagination {page,pageSize,total,pageCount}" }
        "401": { description: Oturum gerekli }
    post:
      tags: [orders]
      summary: Yeni sipariş (40-kolon XLS sözleşmesi; orderCreateSchema)
      responses:
        "200": { description: "message + order" }
        "401": { description: Oturum gerekli }
        "422": { description: Doğrulama hatası }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: cerberus_session
      description: HttpOnly, imzalı ve süreli JWT oturum çerezi
...